Split from_input tests into modular files
